pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "politifact-scraping"
version = "0.1.0"
description = "Scrape content from the PolitiFact website"
readme = "README.md"
license = "CC-BY-4.0"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12",
    "requests>=2.28",
    "fuzzywuzzy>=0.18",
    "python-Levenshtein>=0.21",
    "pymongo>=4.0",
    "python-dotenv>=1.0",
    "numpy>=1.24",
    "pydantic-core>=2.0",
    "toml>=0.10",
]

[project.urls]
Repository = "https://github.com/villarmario/politifact-scraping"

[tool.setuptools.packages.find]
where = ["src"]
